Separate primitives and elements to own files
