+++ board_editor.py
@@ -1,0 +1,54 @@
+#!/usr/bin/env python3
+# -*- coding: UTF-8 -*-
+# file from https://github.com/Nircek/resistorer
+# licensed under MIT license
+
+'''
+MIT License
+
+Copyright (c) 2018-2019 Nircek
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+'''
+
+from elements import Resistor
+
+UNITS = {'R': '\N{OHM SIGN}', 'U': 'V', 'I': 'A'}
+
+
+def get_unit(what):
+    '''get_unit translates the name of variable to corresponding unit of it'''
+    if what in UNITS:
+        return UNITS[what]
+    return ''
+
+
+class CanceledError(Exception):
+    pass
+
+
+def input_resistance(uid, get_float):  # TODO: move it to BoardEditor
+    new_value = None
+    new_value = Resistor.old_r
+    new_value = get_float(
+        'Value of R' + str(uid) + ' [' + get_unit('R') + ']')
+    if new_value is None:
+        raise CanceledError
+    Resistor.old_r = new_value
+    return new_value
