+++ main.py
@@ -34,129 +34,9 @@ import math
 import pickle
 import sys
 
-UNITS = {'R': '\N{OHM SIGN}', 'U': 'V', 'I': 'A'}
-
-
-def get_unit(what):
-    '''get_unit translates the name of variable to corresponding unit of it'''
-    if what in UNITS:
-        return UNITS[what]
-    return ''
-
-
-class Primitive:
-    def __init__(self, r=None):
-        if r is not None:
-            self.ph_r = r
-        self._ph_i, self._ph_u = None, None
-        self.node_a, self.node_b = None, None
-
-    def __repr__(self):
-        return '[' + str(self.ph_r) + ']'
-
-    def __str__(self):
-        return self.__class__.__name__
-
-    @property
-    def components(self):
-        return []
-
-    @property
-    def ph_i(self):
-        return self._ph_i
-
-    @property
-    def ph_u(self):
-        return self._ph_u
-
-    def update_r(self):
-        pass
-
-    def clear_iu(self):
-        self._ph_i = None
-        self._ph_u = None
-
-    @ph_i.setter
-    def ph_i(self, new):
-        if new is None:
-            self.clear_iu()
-            return
-        self._ph_i = new
-        self._ph_u = self.ph_r * new
-        self.update_r()
-
-    @ph_u.setter
-    def ph_u(self, new):
-        if new is None:
-            self.clear_iu()
-            return
-        self._ph_u = new
-        self._ph_i = new / self.ph_r
-        self.update_r()
-
-
-class Series(Primitive):
-    def __init__(self, *args):
-        super().__init__()
-        self.data = args
-        self.ph_u = None
-
-    def __repr__(self):
-        return '+(' + ', '.join(map(repr, self.data)) + ')'
-
-    @property
-    def components(self):
-        return self.data
-
-    @property
-    def ph_r(self):
-        return sum(map(lambda x: x.ph_r, self.data))
-
-    def update_r(self):
-        for ele in self.data:
-            ele.ph_i = self.ph_i
-
-
-class Parallel(Primitive):
-    def __init__(self, *args):
-        super().__init__()
-        self.data = args
-        self.ph_u = None
-
-    def __repr__(self):
-        return ':(' + ', '.join(map(repr, self.data)) + ')'
-
-    @property
-    def components(self):
-        return self.data
-
-    @property
-    def ph_r(self):
-        return 1 / sum(map(lambda x: 1 / x.ph_r, self.data))
-
-    def update_r(self):
-        for ele in self.data:
-            ele.ph_u = self.ph_u
-
-
-class Delta(Primitive):
-    def __init__(self, x, y, z, i):
-        super().__init__()
-        self.components = [x, y, z]
-        self.wiring_type = i
-        self.ph_u = None
-
-    def __repr__(self):
-        return '\N{GREEK CAPITAL LETTER DELTA}(' + \
-               ', '.join(map(repr, self.components + self.wiring_type)) + ')'
-
-    @property
-    def ph_r(self):
-        return {
-            1: self.components[0].ph_r * self.components[1].ph_r,
-            2: self.components[0].ph_r * self.components[2].ph_r,
-            3: self.components[1].ph_r * self.components[2].ph_r
-        }[self.wiring_type] / sum(map(lambda x: x.ph_r, self.components))
+from primitives import Primitive, Series, Parallel, Delta
+from elements import TElement, OElement, APin, BPin, Wire, Resistor
+from board_editor import get_unit, CanceledError
 
 
 class Nodes:
@@ -406,164 +286,6 @@ def pround(x_coord, y_coord, size, is_tel=False):
     x_coord = ((x_coord + 0.5 * size) // size)
     y_coord = ((y_coord + 0.5 * size) // size)
     return Pos(x_coord, y_coord)
-
-
-class Element:
-    def __str__(self):
-        return self.__class__.__name__
-
-    def __init__(self, parent):
-        self.addr = super().__repr__().split('0x')[1][:-1]
-        self.parent = parent
-
-    @property
-    def info(self):
-        return {}
-
-    def __repr__(self):
-        return str(vars(self))
-
-    def onkey(self, event):
-        pass
-
-
-class OElement(Element):
-    def render(self, x_coord, y_coord, size):
-        pass
-
-
-class TElement(Element):
-    def render(self, x_coord, y_coord, size, position):
-        pass
-
-
-class Pin(OElement):
-    def __init__(self, parent, color='black'):
-        super().__init__(parent)
-        self.color = color
-        deleted = []
-        for oel in self.parent.oels:
-            if str(self.parent.oels[oel]) == str(self):
-                deleted += [oel]
-        for oel in deleted:
-            del self.parent.oels[oel]
-
-    def render(self, x_coord, y_coord, size):
-        radius = size * 0.1
-        self.parent.w.create_arc(
-            x_coord - radius, y_coord - radius,
-            x_coord + radius, y_coord + radius,
-            start=0, extent=180,
-            outline=self.color, fill=self.color)
-        self.parent.w.create_arc(x_coord - radius, y_coord - radius,
-                                 x_coord + radius, y_coord + radius,
-                                 start=180, extent=180,
-                                 outline=self.color, fill=self.color)
-        radius *= 2
-        self.parent.w.create_arc(x_coord - radius, y_coord - radius,
-                                 x_coord + radius, y_coord + radius,
-                                 start=0, extent=180,
-                                 outline=self.color, style='arc')
-        self.parent.w.create_arc(x_coord - radius, y_coord - radius,
-                                 x_coord + radius, y_coord + radius,
-                                 start=180, extent=180,
-                                 outline=self.color, style='arc')
-
-
-class APin(Pin):
-    def __init__(self, parent):
-        super().__init__(parent, 'blue')
-
-
-class BPin(Pin):
-    def __init__(self, parent):
-        super().__init__(parent, 'red')
-
-
-class Wire(TElement):
-    def render(self, x_coord, y_coord, size, position):
-        self.parent.w.create_line(x_coord, y_coord, x_coord if position == 1
-                                  else (x_coord + size),
-                                  y_coord if position == 0
-                                  else (y_coord + size))
-
-
-class CanceledError(Exception):
-    pass
-
-
-class Resistor(Primitive, TElement):
-    resistor_i = 1
-    oR = None
-
-    def __init__(self, parent, uid=None):
-        super().__init__(parent)
-        self.parent = parent
-        if uid is None:
-            uid = Resistor.resistor_i
-            Resistor.resistor_i += 1
-        self.uid = uid
-        self.get_r()
-        self.ph_u = None
-
-    def get_r(self):
-        new_value = None
-        new_value = Resistor.oR
-        new_value = self.parent.getFloat(
-            'Value of R' + str(self.uid) + ' [' + get_unit('R') + ']')
-        if new_value is None:
-            raise CanceledError
-        Resistor.oR = new_value
-        self.ph_r = new_value
-
-    @property
-    def info(self):
-        return {'R': self.ph_r, 'U': self.ph_u, 'I': self.ph_i} \
-            if (self.ph_i is not None) and (self.ph_u is not None) \
-            else {'R': self.ph_r}
-
-    def __repr__(self):
-        return '{' + str(self.uid) + '}'
-
-    def render(self, x_coord, y_coord, size, position):
-        if position == 0:
-            self.parent.w.create_line(x_coord, y_coord,
-                                      x_coord + 0.25 * size, y_coord)
-            self.parent.w.create_line(x_coord + 0.75 * size, y_coord,
-                                      x_coord + size, y_coord)
-            self.parent.w.create_line(
-                x_coord + 0.25 * size, y_coord + 0.2 * size,
-                x_coord + 0.75 * size, y_coord + 0.2 * size)
-            self.parent.w.create_line(
-                x_coord + 0.25 * size, y_coord - 0.2 * size,
-                x_coord + 0.75 * size, y_coord - 0.2 * size)
-            self.parent.w.create_line(
-                x_coord + 0.25 * size, y_coord + 0.2 * size,
-                x_coord + 0.25 * size, y_coord - 0.2 * size)
-            self.parent.w.create_line(
-                x_coord + 0.75 * size, y_coord + 0.2 * size,
-                x_coord + 0.75 * size, y_coord - 0.2 * size)
-            self.parent.w.create_text(x_coord + 0.5 * size, y_coord,
-                                      text=str(self.uid))
-        if position == 1:
-            self.parent.w.create_line(x_coord, y_coord,
-                                      x_coord, y_coord + 0.25 * size)
-            self.parent.w.create_line(x_coord, y_coord + 0.75 * size,
-                                      x_coord, y_coord + size)
-            self.parent.w.create_line(
-                x_coord + 0.2 * size, y_coord + 0.25 * size,
-                x_coord + 0.2 * size, y_coord + 0.75 * size)
-            self.parent.w.create_line(
-                x_coord - 0.2 * size, y_coord + 0.25 * size,
-                x_coord - 0.2 * size, y_coord + 0.75 * size)
-            self.parent.w.create_line(
-                x_coord + 0.2 * size, y_coord + 0.25 * size,
-                x_coord - 0.2 * size, y_coord + 0.25 * size)
-            self.parent.w.create_line(
-                x_coord + 0.2 * size, y_coord + 0.75 * size,
-                x_coord - 0.2 * size, y_coord + 0.75 * size)
-            self.parent.w.create_text(
-                x_coord, y_coord + 0.5 * size, text=str(self.uid), angle=270)
 
 
 class Board:
@@ -953,7 +675,7 @@ class Board:
         if event.keysym.upper() == 'O' and event.state == 0b100:  # Ctrl+O
             self.open()
 
-    def get_float(self, msg):
+    def get_float(self, msg):  # TODO: rename it to input_float
         buffer = simpledialog.askfloat('Input', msg,
                                        parent=self.tkroot, minvalue=0.0)
         self.canvas.focus_set()
